+++ lib/pact/matching_rules/jsonpath.rb
@@ -13,7 +13,6 @@ module Pact
       def initialize(path)
         scanner = StringScanner.new(path)
         @path = []
-        bracket_count = 0
         while not scanner.eos?
           if token = scanner.scan(/\$/)
             @path << token
@@ -49,10 +48,10 @@ module Pact
             @path.last << token
           end
         end
+      end
 
-        def to_s
-          path.join
-        end
+      def to_s
+        path.join
       end
     end
   end