+++ spec/lib/pact/matchers/extract_diff_messages_spec.rb
@@ -4,54 +4,54 @@ module Pact
   module Matchers
     describe ExtractDiffMessages do
 
-      subject { ExtractDiffMessages.call(diff) }
+      subject { ExtractDiffMessages.call(diff).join("\n") }
 
       context "<path> with a diff in a Hash" do
         let(:diff) { {a: Difference.new(nil, nil, "There was a difference at <path>") } }
 
-        it { is_expected.to eq "* There was a difference at $.a" }
+        it { is_expected.to eq "There was a difference at $.a" }
       end
 
       context "<parent_path> with a diff in a Hash" do
         let(:diff) { {a: Difference.new(nil, nil, "There was a difference at <parent_path>") } }
 
-        it { is_expected.to eq "* There was a difference at $" }
+        it { is_expected.to eq "There was a difference at $" }
       end
 
       context "<path> with a diff in a nested Hash" do
         let(:diff) { {a: {b: Difference.new(nil, nil, "There was a difference at <path>")}} }
 
-        it { is_expected.to eq "* There was a difference at $.a.b" }
+        it { is_expected.to eq "There was a difference at $.a.b" }
       end
 
       context "<parent_path> with a diff in a nested Hash" do
         let(:diff) { {a: {b: Difference.new(nil, nil, "There was a difference at <parent_path>")}} }
 
-        it { is_expected.to eq "* There was a difference at $.a" }
+        it { is_expected.to eq "There was a difference at $.a" }
       end
 
       context "<path> with a diff in an Array" do
         let(:diff) { [NoDiffAtIndex.new, Difference.new(nil, nil, "There was a difference at <path>")] }
 
-        it { is_expected.to eq "* There was a difference at $[1]" }
+        it { is_expected.to eq "There was a difference at $[1]" }
       end
 
       context "<parent_path> with a diff in an Array" do
         let(:diff) { [NoDiffAtIndex.new, Difference.new(nil, nil, "There was a difference at <parent_path>")] }
 
-        it { is_expected.to eq "* There was a difference at $" }
+        it { is_expected.to eq "There was a difference at $" }
       end
 
       context "<path> with a diff in a nested Array" do
         let(:diff) { [NoDiffAtIndex.new,[NoDiffAtIndex.new, Difference.new(nil, nil, "There was a difference at <path>")]] }
 
-        it { is_expected.to eq "* There was a difference at $[1][1]" }
+        it { is_expected.to eq "There was a difference at $[1][1]" }
       end
 
       context "<parent_path> with a diff in a nested Array" do
         let(:diff) { [NoDiffAtIndex.new,[NoDiffAtIndex.new, Difference.new(nil, nil, "There was a difference at <parent_path>")]] }
 
-        it { is_expected.to eq "* There was a difference at $[1]" }
+        it { is_expected.to eq "There was a difference at $[1]" }
       end
 
       context "when there is a space in the key" do
@@ -59,7 +59,7 @@ module Pact
           {"Foo Bar" => Difference.new(nil, nil, "There was a difference at <path>")}
         end
 
-        it { is_expected.to eq "* There was a difference at $.\"Foo Bar\"" }
+        it { is_expected.to eq "There was a difference at $.\"Foo Bar\"" }
       end
 
       context "with two differences" do
@@ -70,7 +70,7 @@ module Pact
           }
         end
 
-        it { is_expected.to eq "* There was a difference at $.a\n* There was a difference at $.b" }
+        it { is_expected.to eq "There was a difference at $.a\nThere was a difference at $.b" }
       end
 
     end